Initial Commit
Been Developing this for a while now, I had added git to this project before, but there were some absolute file paths left in the project, also had some submodule issues, and forgot to add some licenses for the dependencies, so I decided to start from fresh. So this is the initial commit, currently the FlameUI project can render quads as well as text using batch rendering using the opengl API. Also the font rendering is done using Multi-Signed Distance Fields which are generated using the msdfgen library. There are many bugs still, so I'm working on it.