Fixed Bug.
When adding a non-textured quad, with a textured quad in the same batch, only the texture ids for the textured quad were stored, hence while binding textures, the texture slots were getting messed up, and ended up giving undesired texture on the given quad. Fixed that by adding textureId of 0 to the non-textured quad.